Bugfix: kernel can have last component 0
In the differential case, there are many terms with valuation 0, so even if the first k-1
operators have valuation 0, it can happen that the first k rows of the matrix for a given
term is 0. In that case, the kernel will be e.g. (1, 0 ... 0, , 0), and we should return
None: it is not a valid relation to use to raise the valuation.